refactor(Car, Cars): 각 클래스를 인터페이스와 구현클래스로 추상화 후 관련 코드 변경
변경사항
- Car 클래스를 Car 인터페이스와 GameCar 구현체로 변경
- Cars 클래스를 Cars 인터페이스와 GameCars 구현체로 변경
- Cars record가 class로 변경됨에 따라 차량 cars getter의 네이밍 변경
- 인터페이스화에 따른 RacingGameController와 RacingGameState의 참조 타입 변경
